create database class that all others inherit from
